+++ src/main/java/controllers/FeedController.java
@@ -14,16 +14,15 @@ public class FeedController {
     private FeedManager feedManager;
     private LoginManager loginManager;
     private InOut inOut;
-    private DatabaseManager databaseManager;
+    private DatabaseManager databaseManager = new DatabaseManager();
     private String postsString;
     private HashMap<Integer, Post> postsActionMap = new HashMap<>();
     private String postActionPrompt = "Please select an action for this post: \n"
             + "0 Like the Post";
 
-    public FeedController(InOut inOut, DatabaseManager dbManager, LoginManager loginManager) {
+    public FeedController(InOut inOut, LoginManager loginManager) {
         this.inOut = inOut;
         this.loginManager = loginManager;
-        this.databaseManager = dbManager;
         User currUser = this.loginManager.getCurrUser();
         Post[] posts = this.databaseManager.getAllPosts();
         Feed currUserFeed = new Feed(new ArrayList<>(Arrays.asList(posts)));
@@ -75,12 +74,12 @@ public class FeedController {
 
     private String generateFeedActions(ArrayList<Post> posts) {
         this.postsActionMap = new HashMap<>();
-        StringBuilder postsString = new StringBuilder("Enter a number for a detailed view of a post: ");
+        String postsString = "Enter a number for a detailed view of a post: ";
 
         for (int i=0; i < posts.size(); i++) {
             this.postsActionMap.put(i, posts.get(i));
-            postsString.append("\n").append(i).append(" ").append(posts.get(i).getRecipe().getTitle());
+            postsString += "\n" + i + " " + posts.get(i).getRecipe().getTitle();
         }
-        return postsString.toString();
+        return postsString;
     }
 }
