+++ src/main/java/controllers/LoginController.java
@@ -2,7 +2,6 @@ package controllers;
 
 import entities.InOut;
 import entities.ShellAction;
-import use_cases.DatabaseManager;
 import use_cases.LoginManager;
 
 import java.io.IOException;
@@ -12,10 +11,9 @@ public class LoginController {
     private String loginMessage = "Please enter your login info.";
     private String signUpMessage = "Please signup below:";
     private String welcomeMessage = "Welcome to the Recipe App!";
-    private String welcomeActionPrompt = """
-            Please select an action:
-            0 Sign up
-            1 Login""";
+    private String welcomeActionPrompt = "Please select an action: \n"
+            + "0 Sign up\n"
+            + "1 Login";
     private InOut inOut;
 
     public LoginController(InOut inOut, LoginManager loginManager) {
