+++ src/main/java/controllers/UserProfileController.java
@@ -14,26 +14,21 @@ import use_cases.UserManager;
 
 public class UserProfileController {
     private InOut inOut;
-    private String customizeProfileScreen = "Select which option you'd like to customize:\n" +
-            "0 Username\n" +
-            "1 Password\n" +
-            "2 Bio\n" +
-            "3 Posts\n" +
-            "4 Following list\n";
-    private String otherUserScreen = "Select which action you'd like to take:\n" +
-            "0 Follow user\n" +
-            "1 Browse user's posts\n";
+    private String customizeProfileScreen = "Select which option you'd like to customize:" +
+            "\n0 Username \n1 Password \n2 Bio \n3 Posts \n4 Following list";
+    private String otherUserScreen = "Select which action you'd like to take:" +
+            "\n0 Follow user \n1 Browse user's posts";
     private String userToBrowsePrompt = "Enter the username of the person you'd like to view: ";
     private LoginManager loginManager;
     private UserManager userManager;
     private DatabaseManager databaseManager;
 
     // Constructor
-    public UserProfileController(InOut inOut, DatabaseManager dbManager, LoginManager loginManager) {
+    public UserProfileController(InOut inOut, LoginManager loginManager) {
         this.inOut = inOut;
         this.loginManager = loginManager;
-        this.databaseManager = dbManager;
-        this.userManager = new UserManager(dbManager);
+        this.databaseManager = new DatabaseManager();
+        this.userManager = new UserManager();
     }
 
 
@@ -214,8 +209,8 @@ public class UserProfileController {
      * Display user's posts
      */
     private void runDisplayUserPosts(User user) {
-        PostManager postManager = new PostManager(this.databaseManager);
-        PostController postController = new PostController(this.inOut, this.databaseManager, this.loginManager);
+        PostManager postManager = new PostManager();
+        PostController postController = new PostController(this.inOut, postManager, this.loginManager);
         for(Post post: this.userManager.getUserPosts(user)) {
             // TODO: UPDATE SO THAT POST ID IS THE PARAM FOR DISPLAYPOST
             // postController.displayPost(post);
