+++ src/main/java/entities/Recipe.java
@@ -9,9 +9,9 @@ public class Recipe {
     private String title;
     private String id = UUID.randomUUID().toString();
 
-    public Recipe(String title, ArrayList<Ingredient> ingredients, ArrayList<String> steps) {
-        this.steps = steps;
-        this.ingredients = ingredients;
+    public Recipe(String title) {
+        this.steps = new ArrayList<>();
+        this.ingredients = new ArrayList<>();
         this.title = title;
     }
 
@@ -32,8 +32,6 @@ public class Recipe {
     }
 
     public String getTitle() { return title; }
-
-    public void setTitle(String title) {this.title = title;}
 
     public String getId() {
         return this.id;
