+++ src/test/java/FeedTest.java
@@ -18,11 +18,11 @@ public class FeedTest {
         this.feed = new Feed(new ArrayList<>(Arrays.asList(
                 new Post(UUID.randomUUID().toString(),
                     LocalDateTime.now(),
-                    new Recipe("Stir Fry", new ArrayList<>(), new ArrayList<>()),
+                    new Recipe("Stir Fry"),
                     "Chinese"),
                 new Post(UUID.randomUUID().toString(),
                     LocalDateTime.now(),
-                    new Recipe("Stir Fry", new ArrayList<>(), new ArrayList<>()),
+                    new Recipe("Stir Fry"),
                     "Chinese")
         )));
     }
